requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pillow==9.5.0

